dashboard-financeirozip-main/dashboard-financeiro/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
pydantic==2.9.2
python-multipart==0.0.9
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
openpyxl==3.1.5
anthropic==0.34.2
langchain==0.2.16
langchain-anthropic==0.1.23
python-dotenv==1.0.1
